pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lambdawaker-draw-pattern"
version = "0.1.2"
description = "Drawing utilities for generating geometric patterns and grids using Pillow and aggdraw"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name = "LambdaWalker", email = "lambdawaker@isdavid.com" }
]
keywords = ["graphics", "drawing", "pillow", "aggdraw", "geometry", "grid", "patterns"]
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Graphics",
]
dependencies = [
  "Pillow>=9.0",
  "aggdraw>=1.3.16.post1",
]

[project.urls]
Homepage = "https://github.com/lambdawalker/python_lw_drawing_tools"
Repository = "https://github.com/lambdawalker/python_lw_drawing_tools"
Issues = "https://github.com/lambdawalker/python_lw_drawing_tools/issues"

# Configure setuptools to discover packages (grid, waves, header)
[tool.setuptools.packages.find]
where = ["src"]
include = ["lambdawaker.*"]


